requirements.txt
numpy==1.24.3
langchain==0.1.9
langchain_openai==0.0.7
Flask==3.0.0
chromadb==0.4.15
line-bot-sdk==3.7.0
requests~=2.31.0
pandas~=2.2.2
pygsheets==2.0.6
redis==5.0.1
python-dotenv==1.0.0
openai==1.12.0
tiktoken
langchain-community
scikit-learn~=1.5.1
faiss-cpu
uvicorn~=0.30.1
fastapi~=0.111.0
pydantic~=2.8.2
aiomysql~=0.2.0